requirements.txt
# Core data processing
pandas>=1.3.3
numpy>=1.21.0

# Web framework
flask>=2.0.1
werkzeug>=2.0.1

# Visualization libraries
plotly>=5.3.1
panel>=1.0.0
geopandas>=0.10.0

# Google Cloud services
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# HTTP requests
requests>=2.25.0

# Progress bars
tqdm>=4.60.0

# Data serialization
pickle-mixin>=1.0.2

# Optional: For development and testing
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.8.0
